refactor: 자동 로그인과 다이렉트 기능 분리
